feat: add deployment and service manifests for api-gateway
The commit adds deployment and service manifests for the api-gateway application. The deployment specifies a replica count of 1 and uses the image "muhamedjasim/api-gateway:latest". The deployment also sets environment variables for the AUTH_SVC_URL, PORT, PRODUCT_SVC_URL, and ORDER_SVC_URL. The service is of type NodePort and exposes port 3000.