feat: update routes and request fields in product and order services
The changes in this commit update the routes and request fields in the product and order services. 

In the product service, the "/decrease-stock" route has been removed. 

In the order service, the "OrderId" field in the CreateOrderRequest struct has been renamed to "UserId". 

Additionally, the configuration path for loading environment variables has been updated in the NewConfig function. The path has been changed from "./pkg/config/envs" to "./".